NN Experiment with MNIST data to explore computer vision
The purpose of this paper is to define how neural networks are used to resolve a problem of MNIST handwritten digit classification and to understand how the neurons/nodes in a simple single-hidden layer network have learned to represent features within the input data. We designed a neural network model and then implemented it with various optimization strategies to solve the classification problem. Our multi-layer artificial neural network has an accuracy of 97.20% with test performance.